Atualização - 29/AGO

+++ CollectorQi/Models/ESCL021/DepositosGuardaMaterialItem.cs
@@ -5,6 +5,6 @@ namespace CollectorQi.Models.ESCL021
     public class DepositosGuardaMaterialItem
     {
         public string CodigoItem { get; set; }
-        public int SaldoInfo { get; set; }        
+        public decimal SaldoInfo { get; set; }        
     }
 }
